frontend: Read list of customers from Customer service
Before, the list of customers was hard-coded in the HTML. This meant
that if customers in the database were changed or added, they wouldn't get
reflected in the frontend. It also meant that changing information about
the customer in the frontend could cause a data mismatch when API
requests are made that interact with the customers stored in the
database.

Note that this change negates Will DelHagen's change of changing the
"Amazing Coffee Roasters" customer name to "Kevin". The change will be
properly reintroduced via the database at a later time.